+++ implementations/rust/ockam/ockam/src/lib.rs
@@ -43,7 +43,6 @@ pub use metadata::OckamMessage;
 pub use system::{SystemBuilder, SystemHandler, WorkerSystem};
 pub use unique::unique_with_prefix;
 
-pub mod abac;
 pub mod channel;
 pub mod pipe;
 pub mod pipe2;
@@ -52,6 +51,7 @@ pub mod remote;
 pub mod stream;
 pub mod workers;
 
+pub use ockam_abac as abac;
 pub use ockam_identity as identity;
 
 pub use ockam_core::{
@@ -62,7 +62,6 @@ pub use ockam_core::{
 
 /// Access Control
 pub mod access_control {
-    pub use crate::abac::access_control::*;
     pub use ockam_core::access_control::*;
     pub use ockam_identity::access_control::*;
     pub use ockam_node::access_control::*;
