+++ implementations/rust/ockam/ockam_abac/Cargo.toml
@@ -1,0 +1,44 @@
+[package]
+name = "ockam_abac"
+version = "0.1.0"
+authors = ["Ockam Developers"]
+edition = "2021"
+license = "Apache-2.0"
+homepage = "https://github.com/build-trust/ockam"
+repository = "https://github.com/build-trust/ockam/tree/develop/implementations/rust/ockam/ockam_abac"
+readme = "README.md"
+categories = ["cryptography", "authorization"]
+keywords = ["ockam", "crypto", "encryption", "authorization"]
+description = """Abac is responsible for performing attribute based
+authorization control on messages within an Ockam worker system"""
+publish = true
+rust-version = "1.56.0"
+
+[features]
+default = ["std"]
+
+# Feature (enabled by default): "std" enables functionality expected to
+# be available on a standard platform.
+std = [
+    "ockam_core/std",
+    "ockam_identity/std",
+]
+
+# Feature: "no_std" enables functionality required for platforms
+# without the standard library.
+no_std = [
+    "ockam_core/no_std",
+    "ockam_identity/no_std",
+]
+
+# Feature: "alloc" enables support for heap allocation on "no_std"
+# platforms, requires nightly.
+alloc = [
+    "ockam_core/alloc",
+    "ockam_identity/alloc",
+]
+
+[dependencies]
+ockam_core = { path = "../ockam_core", version = "^0.61.0", default-features = false }
+ockam_identity = { path = "../ockam_identity", version = "^0.55.0", default_features = false }
+serde = { version = "1.0", default-features = false, features = ["derive"] }
