+++ implementations/rust/ockam/ockam_abac/README.md
@@ -1,0 +1,40 @@
+# ockam_abac
+
+[![crate][crate-image]][crate-link]
+[![docs][docs-image]][docs-link]
+[![license][license-image]][license-link]
+[![discuss][discuss-image]][discuss-link]
+
+Ockam is a library for building devices that communicate securely, privately
+and trustfully with cloud services and other devices.
+
+Abac is responsible for performing attribute based authorization
+control on messages within an Ockam worker system.
+
+## Usage
+
+Add this to your `Cargo.toml`:
+
+```
+[dependencies]
+ockam_abac = "0.1.0"
+```
+
+## License
+
+This code is licensed under the terms of the [Apache License 2.0][license-link].
+
+[main-ockam-crate-link]: https://crates.io/crates/ockam
+[ockam-abac-crate-link]: https://crates.io/crates/ockam_abac
+
+[crate-image]: https://img.shields.io/crates/v/ockam_abac.svg
+[crate-link]: https://crates.io/crates/ockam_abac
+
+[docs-image]: https://docs.rs/ockam_abac/badge.svg
+[docs-link]: https://docs.rs/ockam_abac
+
+[license-image]: https://img.shields.io/badge/License-Apache%202.0-green.svg
+[license-link]: https://github.com/build-trust/ockam/blob/HEAD/LICENSE
+
+[discuss-image]: https://img.shields.io/badge/Discuss-Github%20Discussions-ff70b4.svg
+[discuss-link]: https://github.com/build-trust/ockam/discussions
