+++ implementations/rust/ockam/ockam_abac/src/error.rs
