+++ implementations/rust/ockam/ockam_abac/src/lib.rs
@@ -1,21 +1,14 @@
 //! Attribute Based Access Control
 
-pub mod access_control;
 pub mod error;
 
 /// An example abac backend
 pub mod mem;
 
-mod local_info;
-mod metadata;
 mod policy;
 mod traits;
 mod types;
-mod workers;
 
-pub use local_info::*;
-pub use metadata::*;
 pub use policy::*;
 pub use traits::*;
 pub use types::*;
-pub use workers::*;
