+++ implementations/rust/ockam/ockam_examples/example_projects/access_control/examples/04-abac-for-transport-responder.rs
@@ -1,11 +1,11 @@
 // This node starts a tcp listener and an echoer worker.
 // It then runs forever waiting for messages.
 
+use abac_examples::access_control::{AbacUnwrapperWorker, AttributeBasedAccessControl};
 use abac_examples::Echoer;
-use ockam::abac::{
-    self, AbacAttributeStorage, AbacPolicyStorage, AbacUnwrapperWorker, Action, Resource, Subject,
-};
-use ockam::access_control::{AllowedTransport, AttributeBasedAccessControl, LocalOriginOnly};
+
+use ockam::abac::{self, AbacAttributeStorage, AbacPolicyStorage, Action, Resource, Subject};
+use ockam::access_control::{AllowedTransport, LocalOriginOnly};
 use ockam::{Context, Result, TcpTransport, WorkerBuilder, TCP};
 
 #[ockam::node(access_control = "LocalOriginOnly")]
