+++ implementations/rust/ockam/ockam_examples/example_projects/access_control/src/access_control/mod.rs
@@ -1,0 +1,9 @@
+mod access_control;
+mod local_info;
+mod metadata;
+mod workers;
+
+pub use access_control::*;
+pub use local_info::*;
+pub use metadata::*;
+pub use workers::*;
