+++ implementations/rust/ockam/ockam_examples/example_projects/access_control/src/lib.rs
@@ -1,16 +1,14 @@
-mod abac_authorization_worker;
-pub use abac_authorization_worker::*;
-
-mod abac_policy_worker;
-pub use abac_policy_worker::*;
-
-mod authenticated_table_worker;
-pub use authenticated_table_worker::*;
-
-mod echoer;
-pub use echoer::*;
-
-mod hop;
-pub use hop::*;
-
+pub mod access_control;
 pub mod fixtures;
+
+mod abac_authorization_worker;
+mod abac_policy_worker;
+mod authenticated_table_worker;
+mod echoer;
+mod hop;
+
+pub use abac_authorization_worker::*;
+pub use abac_policy_worker::*;
+pub use authenticated_table_worker::*;
+pub use echoer::*;
+pub use hop::*;
